First pass over TwoTowerModel
