+++ merlin_models/tf/models/ranking.py
@@ -6,8 +6,8 @@ from ..blocks.cross import CrossBlock
 from ..blocks.dlrm import DLRMBlock
 from ..blocks.inputs import InputBlock
 from ..blocks.mlp import MLPBlock
-from ..blocks.multi_task import PredictionTasks
 from ..core import Block, Model, ParallelPredictionBlock, PredictionTask
+from .utils import _parse_prediction_tasks
 
 
 def DLRMModel(
@@ -58,17 +58,3 @@ def DCNModel(
 
 def YoutubeDNNRankingModel(schema: Schema) -> Model:
     raise NotImplementedError()
-
-
-def _parse_prediction_tasks(
-    schema: Schema,
-    prediction_tasks: Optional[
-        Union[PredictionTask, List[PredictionTask], ParallelPredictionBlock]
-    ] = None,
-):
-    if not prediction_tasks:
-        prediction_tasks = PredictionTasks(schema)
-    if isinstance(prediction_tasks, (list, tuple)):
-        prediction_tasks = ParallelPredictionBlock(*prediction_tasks)
-
-    return prediction_tasks
