+++ tests/tf/prediction/test_item_prediction.py
@@ -310,23 +310,6 @@ def test_last_item_prediction_task(
 
 
 @pytest.mark.parametrize("run_eagerly", [True, False])
-def test_youtube_dnn_retrieval(
-    sequence_testing_data: SyntheticData,
-    run_eagerly: bool,
-):
-    model = ml.YoutubeDNNRetrievalModel(schema=sequence_testing_data.schema)
-    model.compile(optimizer="adam", run_eagerly=run_eagerly)
-
-    losses = model.fit(sequence_testing_data.tf_dataloader(batch_size=50), epochs=2)
-
-    assert len(losses.epoch) == 2
-    for metric in losses.history.keys():
-        assert type(losses.history[metric]) is list
-    out = model(sequence_testing_data.tf_tensor_dict)
-    assert out.shape[-1] == 51997
-
-
-@pytest.mark.parametrize("run_eagerly", [True, False])
 @pytest.mark.parametrize("ignore_last_batch_on_sample", [True, False])
 def test_retrieval_task_inbatch_default_sampler(
     music_streaming_data: SyntheticData, run_eagerly, ignore_last_batch_on_sample
